Clean-up article scraper. #patch
